Modify the data type of age from float to int.

+++ app/src/main/java/com/example/mlseriesdemonstrator/GA.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class GA implements Serializable {
 
     public int gender;
-    public float age;
+    public int age;
     public enum videoType {
         Child,
         Young,
